+++ src/click/core.py
@@ -19,6 +19,7 @@ from itertools import repeat
 from types import TracebackType
 
 from . import types
+from ._utils import UNSET
 from .exceptions import Abort
 from .exceptions import BadParameter
 from .exceptions import ClickException
@@ -686,14 +687,14 @@ class Context:
             Added the ``call`` parameter.
         """
         if self.default_map is not None:
-            value = self.default_map.get(name)
+            value = self.default_map.get(name, UNSET)
 
             if call and callable(value):
                 return value()
 
             return value
 
-        return None
+        return UNSET
 
     def fail(self, message: str) -> t.NoReturn:
         """Aborts the execution of the program with a specific error
@@ -834,7 +835,7 @@ class Context:
         :rtype: ParameterSource
 
         .. versionchanged:: 8.0
-            Returns ``None`` if the parameter was not provided from any
+            Returns `None` if the parameter was not provided from any
             source.
         """
         return self._parameter_source.get(name)
@@ -863,7 +864,7 @@ class Command:
                             If enabled this will add ``--help`` as argument
                             if no arguments are passed
     :param hidden: hide this command from help outputs.
-    :param deprecated: If ``True`` or non-empty string, issues a message
+    :param deprecated: If `True` or non-empty string, issues a message
                         indicating that the command is deprecated and highlights
                         its deprecation in --help. The message can be customized
                         by using a string as the value.
@@ -1024,7 +1025,7 @@ class Command:
     def get_help_option(self, ctx: Context) -> Option | None:
         """Returns the help option object.
 
-        Skipped if :attr:`add_help_option` is ``False``.
+        Skipped if :attr:`add_help_option` is `False`.
 
         .. versionchanged:: 8.1.8
             The help option is now cached to avoid creating it multiple times.
@@ -1194,7 +1195,7 @@ class Command:
         opts, args, param_order = parser.parse_args(args=args)
 
         for param in iter_params_for_processing(param_order, self.get_params(ctx)):
-            value, args = param.handle_parse_result(ctx, opts, args)
+            _, args = param.handle_parse_result(ctx, opts, args)
 
         if args and not ctx.allow_extra_args and not ctx.resilient_parsing:
             ctx.fail(
@@ -1734,7 +1735,7 @@ class Group(Command):
 
     def get_command(self, ctx: Context, cmd_name: str) -> Command | None:
         """Given a context and a command name, this returns a :class:`Command`
-        object if it exists or returns ``None``.
+        object if it exists or returns `None`.
         """
         return self.commands.get(cmd_name)
 
@@ -2015,14 +2016,16 @@ class Parameter:
     :param is_eager: eager values are processed before non eager ones.  This
                      should not be set for arguments or it will inverse the
                      order of processing.
-    :param envvar: a string or list of strings that are environment variables
-                   that should be checked.
+    :param envvar: environment variable(s) that are used to provide a default value for
+                   this parameter. This can be a string or a sequence of strings. If a
+                   sequence is given, only the first non-empty environment variable is
+                   used for the parameter.
     :param shell_complete: A function that returns custom shell
         completions. Used instead of the param's type completion if
         given. Takes ``ctx, param, incomplete`` and must return a list
         of :class:`~click.shell_completion.CompletionItem` or a list of
         strings.
-    :param deprecated: If ``True`` or non-empty string, issues a message
+    :param deprecated: If `True` or non-empty string, issues a message
                         indicating that the argument is deprecated and highlights
                         its deprecation in --help. The message can be customized
                         by using a string as the value. A deprecated parameter
@@ -2057,7 +2060,7 @@ class Parameter:
 
     .. versionchanged:: 8.0
         Setting a default is no longer required for ``nargs>1``, it will
-        default to ``None``. ``multiple=True`` or ``nargs=-1`` will
+        default to `None`. ``multiple=True`` or ``nargs=-1`` will
         default to ``()``.
 
     .. versionchanged:: 7.1
@@ -2078,7 +2081,7 @@ class Parameter:
         param_decls: cabc.Sequence[str] | None = None,
         type: types.ParamType | t.Any | None = None,
         required: bool = False,
-        default: t.Any | t.Callable[[], t.Any] | None = None,
+        default: t.Any | t.Callable[[], t.Any] | None = UNSET,
         callback: t.Callable[[Context, Parameter, t.Any], t.Any] | None = None,
         nargs: int | None = None,
         multiple: bool = False,
@@ -2127,40 +2130,6 @@ class Parameter:
                     f" type {self.type!r}, but it was {nargs}."
                 )
 
-            # Skip no default or callable default.
-            check_default = default if not callable(default) else None
-
-            if check_default is not None:
-                if multiple:
-                    try:
-                        # Only check the first value against nargs.
-                        check_default = next(_check_iter(check_default), None)
-                    except TypeError:
-                        raise ValueError(
-                            "'default' must be a list when 'multiple' is true."
-                        ) from None
-
-                # Can be None for multiple with empty default.
-                if nargs != 1 and check_default is not None:
-                    try:
-                        _check_iter(check_default)
-                    except TypeError:
-                        if multiple:
-                            message = (
-                                "'default' must be a list of lists when 'multiple' is"
-                                " true and 'nargs' != 1."
-                            )
-                        else:
-                            message = "'default' must be a list when 'nargs' != 1."
-
-                        raise ValueError(message) from None
-
-                    if nargs > 1 and len(check_default) != nargs:
-                        subject = "item length" if multiple else "length"
-                        raise ValueError(
-                            f"'default' {subject} must match nargs={nargs}."
-                        )
-
             if required and deprecated:
                 raise ValueError(
                     f"The {self.param_type_name} '{self.human_readable_name}' "
@@ -2175,6 +2144,13 @@ class Parameter:
         Use :meth:`click.Context.to_info_dict` to traverse the entire
         CLI structure.
 
+        .. versionchanged:: 8.3.0
+            Returns `None` for the :attr:`default` if it is :attr:`UNSET`.
+
+            We explicitly hide the :attr:`UNSET` value to the user, as we choose to
+            make it an implementation detail. And because ``to_info_dict`` has been
+            designed for documentation purposes, we return `None` instead.
+
         .. versionadded:: 8.0
         """
         return {
@@ -2186,7 +2162,7 @@ class Parameter:
             "required": self.required,
             "nargs": self.nargs,
             "multiple": self.multiple,
-            "default": self.default,
+            "default": self.default if self.default is not UNSET else None,
             "envvar": self.envvar,
         }
 
@@ -2254,7 +2230,7 @@ class Parameter:
         """
         value = ctx.lookup_default(self.name, call=False)  # type: ignore
 
-        if value is None:
+        if value is UNSET:
             value = self.default
 
         if call and callable(value):
@@ -2268,20 +2244,39 @@ class Parameter:
     def consume_value(
         self, ctx: Context, opts: cabc.Mapping[str, t.Any]
     ) -> tuple[t.Any, ParameterSource]:
-        value = opts.get(self.name)  # type: ignore
-        source = ParameterSource.COMMANDLINE
+        """Returns the parameter value produced by the parser.
 
-        if value is None:
-            value = self.value_from_envvar(ctx)
-            source = ParameterSource.ENVIRONMENT
+        If the parser did not produce a value from user input, the value is either
+        sourced from the environment variable, the default map, or the parameter's
+        default value. In that order of precedence.
 
-        if value is None:
-            value = ctx.lookup_default(self.name)  # type: ignore
-            source = ParameterSource.DEFAULT_MAP
+        If no value is found, the value is returned as :attr:`UNSET`.
+        """
+        # If the value is set, it means the parser produced a value from user input.
+        value = opts.get(self.name, UNSET)  # type: ignore
+        source = (
+            ParameterSource.COMMANDLINE
+            if value is not UNSET
+            else ParameterSource.DEFAULT
+        )
 
-        if value is None:
-            value = self.get_default(ctx)
-            source = ParameterSource.DEFAULT
+        if value is UNSET:
+            envvar_value = self.value_from_envvar(ctx)
+            if envvar_value is not None:
+                value = envvar_value
+                source = ParameterSource.ENVIRONMENT
+
+        if value is UNSET:
+            default_map_value = ctx.lookup_default(self.name)  # type: ignore
+            if default_map_value is not UNSET:
+                value = default_map_value
+                source = ParameterSource.DEFAULT_MAP
+
+        if value is UNSET:
+            default_value = self.get_default(ctx)
+            if default_value is not UNSET:
+                value = default_value
+                source = ParameterSource.DEFAULT
 
         return value, source
 
@@ -2289,8 +2284,11 @@ class Parameter:
         """Convert and validate a value against the option's
         :attr:`type`, :attr:`multiple`, and :attr:`nargs`.
         """
-        if value is None:
-            return () if self.multiple or self.nargs == -1 else None
+        if value in (None, UNSET):
+            if self.multiple or self.nargs == -1:
+                return ()
+            else:
+                return value
 
         def check_iter(value: t.Any) -> cabc.Iterator[t.Any]:
             try:
@@ -2302,6 +2300,8 @@ class Parameter:
                 raise BadParameter(
                     _("Value must be an iterable."), ctx=ctx, param=self
                 ) from None
+
+        # Define the conversion function based on nargs and type.
 
         if self.nargs == 1 or self.type.is_composite:
 
@@ -2337,7 +2337,15 @@ class Parameter:
         return convert(value)
 
     def value_is_missing(self, value: t.Any) -> bool:
-        if value is None:
+        """
+        A value is considered missing if:
+
+        - it is :attr:`UNSET`,
+        - or if it is an empty sequence while the parameter is suppose to have
+          non-single value (i.e. :attr:`nargs` is not ``1`` or :attr:`multiple` is
+          set).
+        """
+        if value is UNSET:
             return True
 
         if (self.nargs != 1 or self.multiple) and value == ():
@@ -2346,6 +2354,22 @@ class Parameter:
         return False
 
     def process_value(self, ctx: Context, value: t.Any) -> t.Any:
+        """Process the value of this parameter:
+
+        1. Type cast the value using :meth:`type_cast_value`.
+        2. Check if the value is missing (see: :meth:`value_is_missing`), and raise
+           :exc:`MissingParameter` if it is required.
+        3. If a :attr:`callback` is set, call it to have the value replaced buy the
+           result of the callback.
+
+        .. versionchanged:: 8.3
+
+            The :attr:`callback` is now receiving the :attr:`UNSET` sentinel if the
+            parameter is not given by the CLI user.
+
+            The callback is receiving the value as-is, which let the developer decide
+            how to handle the different cases of `None`, ``UNSET``, or any other value.
+        """
         value = self.type_cast_value(ctx, value)
 
         if self.required and self.value_is_missing(value):
@@ -2357,60 +2381,88 @@ class Parameter:
         return value
 
     def resolve_envvar_value(self, ctx: Context) -> str | None:
-        if self.envvar is None:
+        """Returns the value found in the environment variable(s) attached to this
+        parameter.
+
+        Environment variables values are `always returned as strings
+        <https://docs.python.org/3/library/os.html#os.environ>`_.
+
+        This method returns `None` if:
+
+        - the :attr:`envvar` property is not set on the :class:`Parameter`,
+        - the environment variable is not found in the environment,
+        - the variable is found in the environment but its value is empty (i.e. the
+          environment variable is present but has an empty string).
+
+        If :attr:`envvar` is setup with multiple environment variables,
+        then only the first non-empty value is returned.
+
+        .. caution::
+
+            The raw value extracted from the environment is not normalized and is
+            returned as-is. Any normalization or reconciliation is performed later by
+            the :class:`Parameter`'s :attr:`type`.
+        """
+        if not self.envvar:
             return None
 
-        if isinstance(self.envvar, str):
-            rv = os.environ.get(self.envvar)
+        envvars = [self.envvar] if isinstance(self.envvar, str) else self.envvar
+        for envvar in envvars:
+            raw_value = os.environ.get(envvar)
 
-            if rv:
-                return rv
-        else:
-            for envvar in self.envvar:
-                rv = os.environ.get(envvar)
-
-                if rv:
-                    return rv
+            # Return the first non-empty value of the list of environment variables.
+            if raw_value:
+                return raw_value
+            # Else, absence of value is interpreted as an environment variable that is
+            # not set, so proceed to the next one.
 
         return None
 
-    def value_from_envvar(self, ctx: Context) -> t.Any | None:
-        rv: t.Any | None = self.resolve_envvar_value(ctx)
+    def value_from_envvar(self, ctx: Context) -> str | cabc.Sequence[str] | None:
+        """Process the raw environment variable string for this parameter.
 
-        if rv is not None and self.nargs != 1:
-            rv = self.type.split_envvar_value(rv)
+        Returns the string as-is or splits it into a sequence of strings if the
+        parameter is expecting multiple values (i.e. its :attr:`nargs` property is set
+        to a value other than ``1``).
+        """
+        raw_value = self.resolve_envvar_value(ctx)
 
-        return rv
+        if raw_value is not None and self.nargs != 1:
+            return self.type.split_envvar_value(raw_value)
+
+        return raw_value
 
     def handle_parse_result(
         self, ctx: Context, opts: cabc.Mapping[str, t.Any], args: list[str]
     ) -> tuple[t.Any, list[str]]:
+        """Process the value produced by the parser from user input."""
         with augment_usage_errors(ctx, param=self):
             value, source = self.consume_value(ctx, opts)
 
-            if (
-                self.deprecated
-                and value is not None
-                and source
-                not in (
-                    ParameterSource.DEFAULT,
-                    ParameterSource.DEFAULT_MAP,
-                )
-            ):
-                extra_message = (
-                    f" {self.deprecated}" if isinstance(self.deprecated, str) else ""
-                )
-                message = _(
-                    "DeprecationWarning: The {param_type} {name!r} is deprecated."
-                    "{extra_message}"
-                ).format(
-                    param_type=self.param_type_name,
-                    name=self.human_readable_name,
-                    extra_message=extra_message,
-                )
-                echo(style(message, fg="red"), err=True)
-
             ctx.set_parameter_source(self.name, source)  # type: ignore
+
+            # Only try to process the value if it is not coming from the defaults.
+            # Defaults set by the users are under their control and so are left
+            # untouched as Python values: these values does not need to be processed.
+            if source not in (ParameterSource.DEFAULT, ParameterSource.DEFAULT_MAP):
+                # If the parameter is deprecated, we want to warn the user about it.
+                # If the value is UNSET there is no need to warn the user of a
+                # parameter that has not been invoked.
+                if self.deprecated and value is not UNSET:
+                    extra_message = (
+                        f" {self.deprecated}"
+                        if isinstance(self.deprecated, str)
+                        else ""
+                    )
+                    message = _(
+                        "DeprecationWarning: The {param_type} {name!r} is deprecated."
+                        "{extra_message}"
+                    ).format(
+                        param_type=self.param_type_name,
+                        name=self.human_readable_name,
+                        extra_message=extra_message,
+                    )
+                    echo(style(message, fg="red"), err=True)
 
             try:
                 value = self.process_value(ctx, value)
@@ -2418,10 +2470,19 @@ class Parameter:
                 if not ctx.resilient_parsing:
                     raise
 
-                value = None
+                # Ignore values that we have a hard time processing.
+                value = UNSET
 
-        if self.expose_value:
-            ctx.params[self.name] = value  # type: ignore
+        if self.expose_value and self.name not in ctx.params:
+            # Click is logically enforcing that the name is None if the parameter is
+            # not exposed. We still assert it here to please the type checker.
+            assert self.name is not None, (
+                f"{self!r} parameter's name should not be None when exposing value."
+            )
+            # Normalize unset values to None, as we're about to pass them to the
+            # command function and move them to the pure-Python realm of user-written
+            # code.
+            ctx.params[self.name] = value if value is not UNSET else None
 
         return value, args
 
@@ -2470,25 +2531,25 @@ class Option(Parameter):
 
     :param show_default: Show the default value for this option in its
         help text. Values are not shown by default, unless
-        :attr:`Context.show_default` is ``True``. If this value is a
+        :attr:`Context.show_default` is `True`. If this value is a
         string, it shows that string in parentheses instead of the
         actual value. This is particularly useful for dynamic options.
         For single option boolean flags, the default remains hidden if
-        its value is ``False``.
+        its value is `False`.
     :param show_envvar: Controls if an environment variable should be
         shown on the help page and error messages.
         Normally, environment variables are not shown.
-    :param prompt: If set to ``True`` or a non empty string then the
-        user will be prompted for input. If set to ``True`` the prompt
+    :param prompt: If set to `True` or a non empty string then the
+        user will be prompted for input. If set to `True` the prompt
         will be the option name capitalized. A deprecated option cannot be
         prompted.
     :param confirmation_prompt: Prompt a second time to confirm the
         value if it was prompted for. Can be set to a string instead of
-        ``True`` to customize the message.
-    :param prompt_required: If set to ``False``, the user will be
+        `True` to customize the message.
+    :param prompt_required: If set to `False`, the user will be
         prompted for input only when the option was specified as a flag
         without a value.
-    :param hide_input: If this is ``True`` then the input on the prompt
+    :param hide_input: If this is `True` then the input on the prompt
         will be hidden from the user. This is useful for password input.
     :param is_flag: forces this option to act as a flag.  The default is
                     auto detection.
@@ -2522,7 +2583,7 @@ class Option(Parameter):
 
     .. versionchanged:: 8.1
         The default of a single option boolean flag is not shown if the
-        default value is ``False``.
+        default value is `False`.
 
     .. versionchanged:: 8.0.1
         ``type`` is detected from ``flag_value`` if given.
@@ -2539,7 +2600,7 @@ class Option(Parameter):
         prompt_required: bool = True,
         hide_input: bool = False,
         is_flag: bool | None = None,
-        flag_value: t.Any | None = None,
+        flag_value: t.Any = UNSET,
         multiple: bool = False,
         count: bool = False,
         allow_from_autoenv: bool = True,
@@ -2582,46 +2643,47 @@ class Option(Parameter):
         self.hide_input = hide_input
         self.hidden = hidden
 
-        # If prompt is enabled but not required, then the option can be
-        # used as a flag to indicate using prompt or flag_value.
+        # The _flag_needs_value property tells the parser that this option is a flag
+        # that cannot be used standalone and needs a value. With this information, the
+        # parser can determine whether to consider the next user-provided argument in
+        # the CLI as a value for this flag or as a new option.
+        # If prompt is enabled but not required, then it opens the possibility for the
+        # option to gets its value from the user.
         self._flag_needs_value = self.prompt is not None and not self.prompt_required
 
+        # Auto-detect if this is a flag or not.
         if is_flag is None:
             # Implicitly a flag because flag_value was set.
-            if flag_value is not None:
+            if flag_value is not UNSET:
                 is_flag = True
             # Not a flag, but when used as a flag it shows a prompt.
             elif self._flag_needs_value:
                 is_flag = False
-            # Implicitly a flag because flag options were given.
+            # Implicitly a flag because secondary options names were given.
             elif self.secondary_opts:
                 is_flag = True
+        # The option is explicitly not a flag. But we do not know yet if it needs a
+        # value or not. So we look at the default value to determine it.
         elif is_flag is False and not self._flag_needs_value:
-            # Not a flag, and prompt is not enabled, can be used as a
-            # flag if flag_value is set.
-            self._flag_needs_value = flag_value is not None
-
-        self.default: t.Any | t.Callable[[], t.Any]
+            self._flag_needs_value = self.default is UNSET
 
         if is_flag:
             # Set missing default for flags if not explicitly required or prompted.
-            if self.default is None and not self.required and not self.prompt:
+            if self.default is UNSET and not self.required and not self.prompt:
                 if multiple:
                     self.default = ()
+
+            # Auto-detect the type of the flag based on the flag_value.
+            if type is None:
+                # A flag without a flag_value is a boolean flag.
+                if flag_value is UNSET:
+                    self.type = types.BoolParamType()
+                # If the flag value is a boolean, use BoolParamType.
+                elif isinstance(flag_value, bool):
+                    self.type = types.BoolParamType()
+                # Otherwise, guess the type from the flag value.
                 else:
-                    self.default = False
-
-            if flag_value is None:
-                # A boolean flag presence in the command line is enough to set
-                # the value: to the default if it is not blank, or to True
-                # otherwise.
-                flag_value = self.default if self.default else True
-
-        self.type: types.ParamType
-        if is_flag and type is None:
-            # Re-guess the type from the flag value instead of the
-            # default.
-            self.type = types.convert_type(None, flag_value)
+                    self.type = types.convert_type(None, flag_value)
 
         self.is_flag: bool = bool(is_flag)
         self.is_bool_flag: bool = bool(
@@ -2629,12 +2691,24 @@ class Option(Parameter):
         )
         self.flag_value: t.Any = flag_value
 
-        # Counting
+        # Set boolean flag default to False if unset and not required.
+        if self.is_bool_flag:
+            if self.default is UNSET and not self.required:
+                self.default = False
+
+        # Set the default flag_value if it is not set.
+        if self.flag_value is UNSET:
+            if self.is_flag:
+                self.flag_value = True
+            else:
+                self.flag_value = None
+
+        # Counting.
         self.count = count
         if count:
             if type is None:
                 self.type = types.IntRange(min=0)
-            if self.default is None:
+            if self.default is UNSET:
                 self.default = 0
 
         self.allow_from_autoenv = allow_from_autoenv
@@ -2649,9 +2723,6 @@ class Option(Parameter):
 
             if self.nargs == -1:
                 raise TypeError("nargs=-1 is not supported for options.")
-
-            if self.prompt and self.is_flag and not self.is_bool_flag:
-                raise TypeError("'prompt' is not valid for non-boolean flag.")
 
             if not self.is_bool_flag and self.secondary_opts:
                 raise TypeError("Secondary flag is not valid for non-boolean flag.")
@@ -2669,12 +2740,20 @@ class Option(Parameter):
                     raise TypeError("'count' is not valid with 'is_flag'.")
 
     def to_info_dict(self) -> dict[str, t.Any]:
+        """
+        .. versionchanged:: 8.3.0
+            Returns `None` for the :attr:`flag_value` if it is :attr:`UNSET`.
+
+            We explicitly hide the :attr:`UNSET` value to the user, as we choose to
+            make it an implementation detail. And because ``to_info_dict`` has been
+            designed for documentation purposes, we return `None` instead.
+        """
         info_dict = super().to_info_dict()
         info_dict.update(
             help=self.help,
             prompt=self.prompt,
             is_flag=self.is_flag,
-            flag_value=self.flag_value,
+            flag_value=self.flag_value if self.flag_value is not UNSET else None,
             count=self.count,
             hidden=self.hidden,
         )
@@ -2867,7 +2946,9 @@ class Option(Parameter):
         elif ctx.show_default is not None:
             show_default = ctx.show_default
 
-        if show_default_is_str or (show_default and (default_value is not None)):
+        if show_default_is_str or (
+            show_default and (default_value not in (None, UNSET))
+        ):
             if show_default_is_str:
                 default_string = f"({self.show_default})"
             elif isinstance(default_value, (list, tuple)):
@@ -2907,32 +2988,6 @@ class Option(Parameter):
 
         return extra
 
-    @t.overload
-    def get_default(
-        self, ctx: Context, call: t.Literal[True] = True
-    ) -> t.Any | None: ...
-
-    @t.overload
-    def get_default(
-        self, ctx: Context, call: bool = ...
-    ) -> t.Any | t.Callable[[], t.Any] | None: ...
-
-    def get_default(
-        self, ctx: Context, call: bool = True
-    ) -> t.Any | t.Callable[[], t.Any] | None:
-        # If we're a non boolean flag our default is more complex because
-        # we need to look at all flags in the same group to figure out
-        # if we're the default one in which case we return the flag
-        # value as default.
-        if self.is_flag and not self.is_bool_flag:
-            for param in ctx.command.params:
-                if param.name == self.name and param.default is not None:
-                    return t.cast(Option, param).default
-
-            return None
-
-        return super().get_default(ctx, call=call)
-
     def prompt_for_value(self, ctx: Context) -> t.Any:
         """This is an alternative flow that can be activated in the full
         value processing if a value does not exist.  It will prompt the
@@ -2941,12 +2996,18 @@ class Option(Parameter):
         """
         assert self.prompt is not None
 
-        # Calculate the default before prompting anything to be stable.
+        # Calculate the default before prompting anything to lock in the value before
+        # attempting any user interaction.
         default = self.get_default(ctx)
 
-        # If this is a prompt for a flag we need to handle this
-        # differently.
+        # A boolean flag can use a simplified [y/n] confirmation prompt.
         if self.is_bool_flag:
+            # If we have no boolean default, we force the user to explicitly provide
+            # one.
+            if default in (UNSET, None):
+                default = None
+            else:
+                default = bool(default)
             return confirm(self.prompt, default)
 
         # If show_default is set to True/False, provide this to `prompt` as well. For
@@ -2957,7 +3018,9 @@ class Option(Parameter):
 
         return prompt(
             self.prompt,
-            default=default,
+            # Use `None` to inform the prompt() function to reiterate until a valid
+            # value is provided by the user if we have no default.
+            default=None if default is UNSET else default,
             type=self.type,
             hide_input=self.hide_input,
             show_choices=self.show_choices,
@@ -2967,22 +3030,17 @@ class Option(Parameter):
         )
 
     def resolve_envvar_value(self, ctx: Context) -> str | None:
-        """Find which environment variable to read for this option and return
-        its value.
-
-        Returns the value of the environment variable if it exists, or ``None``
-        if it does not.
-
-        .. caution::
-
-            The raw value extracted from the environment is not normalized and
-            is returned as-is. Any normalization or reconciation with the
-            option's type should happen later.
+        """:class:`Option` resolves its environment variable the same way as
+        :func:`Parameter.resolve_envvar_value`, but it also supports
+        :attr:`Context.auto_envvar_prefix`. If we could not find an environment from
+        the :attr:`envvar` property, we fallback on :attr:`Context.auto_envvar_prefix`
+        to build dynamiccaly the environment variable name using the
+        :python:`{ctx.auto_envvar_prefix}_{self.name.upper()}` template.
         """
-        rv = super().resolve_envvar_value(ctx)
+        raw_value = super().resolve_envvar_value(ctx)
 
-        if rv is not None:
-            return rv
+        if raw_value is not None:
+            return raw_value
 
         if (
             self.allow_from_autoenv
@@ -2990,82 +3048,111 @@ class Option(Parameter):
             and self.name is not None
         ):
             envvar = f"{ctx.auto_envvar_prefix}_{self.name.upper()}"
-            rv = os.environ.get(envvar)
+            raw_value = os.environ.get(envvar)
 
-            if rv:
-                return rv
+            # Return the first non-empty value of the list of environment variables.
+            if raw_value:
+                return raw_value
+            # Else, absence of value is interpreted as an environment variable that is
+            # not set, so proceed to the next one.
 
         return None
 
-    def value_from_envvar(self, ctx: Context) -> t.Any | None:
-        """Normalize the value from the environment variable, if it exists."""
-        rv: str | None = self.resolve_envvar_value(ctx)
+    def value_from_envvar(self, ctx: Context) -> t.Any:
+        """
+        For :class:`Option`, this method processes the raw environment variable string
+        the same way as :func:`Parameter.value_from_envvar` does.
 
-        if rv is None:
+        But in the case of non-boolean flags, the value is analyzed to determine if the
+        flag is activated or not, and returns a boolean of its activation, or the
+        :attr:`flag_value` if the latter is set.
+
+        This method also takes care of repeated options (i.e. options with
+        :attr:`multiple` set to `True`).
+        """
+        raw_value = self.resolve_envvar_value(ctx)
+
+        # Absent environment variable or an empty string is interpreted as unset.
+        if raw_value is None:
             return None
 
         # Non-boolean flags are more liberal in what they accept. But a flag being a
-        # flag, its envvar value still needs to analyzed to determine if the flag is
+        # flag, its envvar value still needs to be analyzed to determine if the flag is
         # activated or not.
         if self.is_flag and not self.is_bool_flag:
             # If the flag_value is set and match the envvar value, return it
             # directly.
-            if self.flag_value is not None and rv == self.flag_value:
+            if self.flag_value is not UNSET and raw_value == self.flag_value:
                 return self.flag_value
             # Analyze the envvar value as a boolean to know if the flag is
             # activated or not.
-            return types.BoolParamType.str_to_bool(rv)
+            return types.BoolParamType.str_to_bool(raw_value)
 
         # Split the envvar value if it is allowed to be repeated.
         value_depth = (self.nargs != 1) + bool(self.multiple)
         if value_depth > 0:
-            multi_rv = self.type.split_envvar_value(rv)
+            multi_rv = self.type.split_envvar_value(raw_value)
             if self.multiple and self.nargs != 1:
                 multi_rv = batch(multi_rv, self.nargs)  # type: ignore[assignment]
 
             return multi_rv
 
-        return rv
+        return raw_value
 
     def consume_value(
         self, ctx: Context, opts: cabc.Mapping[str, Parameter]
     ) -> tuple[t.Any, ParameterSource]:
+        """
+        For :class:`Option`, the value can be collected from an interactive prompt if
+        the option is a flag that needs a value (and the :attr:`prompt` property is
+        set).
+
+        Additionally, this method handles flag option that are activated without a
+        value, in which case the :attr:`flag_value` is returned.
+        """
         value, source = super().consume_value(ctx, opts)
 
-        # The parser will emit a sentinel value if the option can be
-        # given as a flag without a value. This is different from None
-        # to distinguish from the flag not being given at all.
+        # The parser will emit a sentinel value if the option is allowed to as a flag
+        # without a value.
         if value is _flag_needs_value:
+            # If the option allows for a prompt, we start an interaction with the user.
             if self.prompt is not None and not ctx.resilient_parsing:
                 value = self.prompt_for_value(ctx)
                 source = ParameterSource.PROMPT
+            # Else the flag takes its flag_value as value.
             else:
                 value = self.flag_value
                 source = ParameterSource.COMMANDLINE
 
-        # A flag which is activated and has a flag_value set, should returns
-        # the latter, unless the value comes from the explicitly sets default.
+        # A flag which is activated always returns the flag value, unless the value
+        # comes from the explicitly sets default.
         elif (
             self.is_flag
             and value is True
             and not self.is_bool_flag
-            and self.flag_value is not None
-            and source is not ParameterSource.DEFAULT
+            and source not in (ParameterSource.DEFAULT, ParameterSource.DEFAULT_MAP)
         ):
             value = self.flag_value
 
+        # Re-interpret a multiple option which has been sent as-is by the parser.
+        # Here we replace each occurrence of value-less flags (marked by the
+        # _flag_needs_value sentinel) with the flag_value.
         elif (
             self.multiple
-            and value is not None
+            and value is not UNSET
+            and source not in (ParameterSource.DEFAULT, ParameterSource.DEFAULT_MAP)
             and any(v is _flag_needs_value for v in value)
         ):
             value = [self.flag_value if v is _flag_needs_value else v for v in value]
             source = ParameterSource.COMMANDLINE
 
-        # The value wasn't set, or used the param's default, prompt if
-        # prompting is enabled.
+        # The value wasn't set, or used the param's default, prompt for one to the user
+        # if prompting is enabled.
         elif (
-            source in {None, ParameterSource.DEFAULT}
+            (
+                value is UNSET
+                or source in (ParameterSource.DEFAULT, ParameterSource.DEFAULT_MAP)
+            )
             and self.prompt is not None
             and (self.required or self.prompt_required)
             and not ctx.resilient_parsing
@@ -3074,6 +3161,14 @@ class Option(Parameter):
             source = ParameterSource.PROMPT
 
         return value, source
+
+    def type_cast_value(self, ctx: Context, value: t.Any) -> t.Any:
+        if self.is_flag and not self.required:
+            if value is UNSET:
+                if self.is_bool_flag:
+                    # If the flag is a boolean flag, we return False if it is not set.
+                    value = False
+        return super().type_cast_value(ctx, value)
 
 
 class Argument(Parameter):
@@ -3092,20 +3187,20 @@ class Argument(Parameter):
         required: bool | None = None,
         **attrs: t.Any,
     ) -> None:
+        # Auto-detect the requirement status of the argument if not explicitly set.
         if required is None:
-            if attrs.get("default") is not None:
-                required = False
-            else:
+            # The argument gets automatically required if it has no explicit default
+            # value set and is setup to match at least one value.
+            if attrs.get("default", UNSET) is UNSET:
                 required = attrs.get("nargs", 1) > 0
+            # If the argument has a default value, it is not required.
+            else:
+                required = False
 
         if "multiple" in attrs:
             raise TypeError("__init__() got an unexpected keyword argument 'multiple'.")
 
         super().__init__(param_decls, required=required, **attrs)
-
-        if __debug__:
-            if self.default is not None and self.nargs == -1:
-                raise TypeError("'default' is not supported for nargs=-1.")
 
     @property
     def human_readable_name(self) -> str: